Split task drill-in layout hook
